# openapi.yaml
openapi: 3.0.0
info:
  title: Pet Adoption System API
  description: API for user registration, login, adopter preferences, shelter pet
    management, and recommendations in the Pet Adoption System.
  version: 5.0.0
  contact:
    name: PRJ666 - Group 1 - 'Pet Adoption System' Team
    email: support@petadoption.com
servers:
  - url: ${openapi.petAdoptionSystem.base-path}
    description: Configurable base path for the API
...
paths:
  /user/register:
    post:
      tags:
        - user
      summary: Register a new user
      description: Registers a new user with an email, password, and account type
        (Pet Adopter or Pet Shelter).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: adopter@example.com
                password:
                  type: string
                  example: P@ssw0rd
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        '201':
          description: User registered successfully
          $ref: '#/components/responses/GenericApiResponse'
        '400':
          description: Email is already taken
          $ref: '#/components/responses/GenericApiResponse'
  /user/login:
    post:
      tags:
        - user
      summary: Login a user
      description: Logs in a user using email and password, and returns a user role and a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: adopter@example.com
                password:
                  type: string
                  example: P@ssw0rd
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: object
                        properties:
                          role:
                            $ref: '#/components/schemas/Role'
                          profileSet:
                            type: boolean
                            example: true
                          token:
                            type: string
                            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid email or password
          $ref: '#/components/responses/GenericApiResponse'
  /user/preferences:
    post:
      tags:
        - user
      summary: Add preferences for pet attributes grouped by attribute groups
      security:
        - BearerAuth: [ ]
      description: >
        Submit a set of preferences for different attribute groups. Each key in the JSON object
        represents an attribute group name, and its corresponding value is a list of selected preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
                description: >
                  A map where each key is the name of an attribute group, and the value is a list of
                  selected preferences within that group.
                example:
                  "Physical Attributes": ["Brown", "Black"]
                  "Behavioral Attributes": ["Active", "Calm"]
                  "Health Attributes": ["Vaccinated", "Neutered"]
      responses:
        '201':
          description: Preferences added successfully
          $ref: '#/components/responses/GenericApiResponse'
  /user/upload-image:
    post:
      tags:
        - user
      summary: Upload an image for a user
      security:
        - BearerAuth: [ ]
      description: Uploads a profile image for a user.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
                - image
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: string
                        example: /images/someImage.jpg
        '400':
          description: Invalid request or missing parameters
          $ref: '#/components/responses/GenericApiResponse'
  /user/get-profile:
    get:
      tags:
        - user
      summary: Retrieve user profile information
      security:
        - BearerAuth: [ ]
      description: Fetches the profile details of the currently authenticated user
        (adopter or shelter).
      responses:
        '200':
          description: Profile details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: object
                        properties:
                          id:
                            type: string
                            example: "12ad3fa..."
                          email:
                            type: string
                            example: user@example.com
                          role:
                            $ref: '#/components/schemas/Role'
                          name:
                            type: string
                            example: John Doe
                          phone:
                            type: string
                            example: 123-456-7890
                          address:
                            type: string
                            example: 123 Pet Lane
                          imageUrl:
                            type: string
                            example: http://some.site/images/img.jpg
        '401':
          description: Unauthorized access
          $ref: '#/components/responses/GenericApiResponse'
  /user/update-profile:
    put:
      tags:
        - user
      summary: Update a user profile
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John Doe
                phone:
                  type: string
                  example: 123-456-7890
                address:
                  type: string
                  example: 123 Pet Lane
                imageUrl:
                  type: string
                  example: http://some.site/images/image.jpg
      responses:
        '200':
          description: User contacts updated successfully
          $ref: '#/components/responses/GenericApiResponse'
  /user/delete-profile/{id}:
    delete:
      tags:
        - user
      summary: Delete a user profile
      description: Deletes the user profile with the specified ID. Requires authentication and authorization.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          description: The userId for the profile that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User profile successfully deleted.
          $ref: '#/components/responses/GenericApiResponse'
        '400':
          description: Invalid username supplied
          $ref: '#/components/responses/GenericApiResponse'
        '401':
          description: Unauthorized access.
          $ref: '#/components/responses/GenericApiResponse'
        '403':
          description: Forbidden. User does not have permission to delete this profile
          $ref: '#/components/responses/GenericApiResponse'
        '404':
          description: User not found
          $ref: '#/components/responses/GenericApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but helps to specify the token type
  responses:
    GenericApiResponse:
      description: Generic API response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    Role:
      type: string
      enum:
        - adopter
        - shelter
      example: adopter
...